+++ src/models/enums.ts
@@ -1,3 +1,5 @@
+import { RemoteSettings } from '../remote/enums';
+
 export enum StandardSettings {
   ElementAdjustments = 'elementAdjustments',
   FavoriteColors = 'favoriteColors',
@@ -19,7 +21,7 @@ export enum AffectedSettings {
   TabActiveBorder = 'affectTabActiveBorder',
 }
 
-export type AllSettings = StandardSettings | AffectedSettings;
+export type AllSettings = StandardSettings | AffectedSettings | RemoteSettings;
 
 export enum Commands {
   resetColors = 'peacock.resetColors',