+++ src/remote/constants.ts
@@ -1,7 +1,0 @@
-import { extensionShortName } from '../models';
-
-export const peacockRemoteMementos = {
-  remoteWslColor: `${extensionShortName}.remote.wsl.color`,
-  remoteSshColor: `${extensionShortName}.remote.ssh.color`,
-  remoteContainersColor: `${extensionShortName}.remote.containers.color`,
-};